miner/requirements.txt
asyncio==3.4.3
black==25.1.0
click==8.1.8
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
websockets==12.0
